separate native calls into different files
